+++ panel/js/urls.js
@@ -22,7 +22,8 @@ var SOURCES = [
   { key: 'twitch',    label: 'Twitch',    re: /twitch\.tv/ },
   { key: 'facebook',  label: 'Facebook',  re: /(?:facebook\.com|fb\.watch)/ },
   { key: 'dailymotion', label: 'Dailymotion', re: /dailymotion\.com/ },
-  { key: 'streamable', label: 'Streamable', re: /streamable\.com/ }
+  { key: 'streamable', label: 'Streamable', re: /streamable\.com/ },
+  { key: 'flow',      label: 'Flow',      re: /labs\.google\/fx\/tools\/flow/ }
 ];
 
 function source(url) {